Correct double vs long for check errors.

+++ src/main/java/com/lazerycode/jmeter/checker/ResultChecker.java
@@ -32,7 +32,7 @@ public class ResultChecker {
                     key, "throughput");
 
             // Check errors
-            double percentErrors = aggregatedResponses.getDuration().getErrorsCount() /
+            double percentErrors = ((double) aggregatedResponses.getDuration().getErrorsCount()) /
                     (aggregatedResponses.getDuration().getErrorsCount() + aggregatedResponses.getDuration().getSuccessCount());
             check &= checkValue(checkResult.getErrors(), percentErrors,
                     key, "errors");
